Only redirect to last used channel when going to app index
Transitioning from the application `afterModel` was triggering the
`model` hook for a second time, which created the same space twice
and also created the sockethub event listeners twice.
